+++ config/api_router.py
@@ -1,7 +1,11 @@
 from django.conf import settings
 from rest_framework.routers import DefaultRouter, SimpleRouter
 
-from architecture_tool_django.modeling.api.views import SchemaViewSet
+from architecture_tool_django.modeling.api.views import (
+    EdgetypeViewSet,
+    NodetypeViewSet,
+    SchemaViewSet,
+)
 
 # from architecture_tool_django.users.api.views import UserViewSet
 
@@ -12,6 +16,8 @@ else:
 
 # router.register("users", UserViewSet)
 router.register("schemas", SchemaViewSet)
+router.register("nodetypes", NodetypeViewSet)
+router.register("edgetypes", EdgetypeViewSet)
 
 app_name = "api"
 urlpatterns = router.urls